init project
init backend mongodb and frontend
the backend is confirgured with authorisation - access token and refresh token - with all the required db models - has been tested via postman. also there is and api for confirming a user - after registering the user gets an email with a url with an id - if you send this to /api/auth/confirm the field of isConfirmed in users will be set to true.
in the frontend i add a login and register page, routing, context, react-query and a navbar for the project.